Move histogram validation code to model/histogram
